First REST version of the server
Provides GET and POST verbs for the waypoint and mission resources
without user authentication. It also provides an utility for admins
to create waypoints and missions.

It removes all the old RPC calls and the admin HTML interface. It
changes the namespace to keep backwards compatibility.

The version for this instance is dev.